backend/requirements.txt
# OLD ONE
# fastapi==0.104.1
# uvicorn==0.24.0
# python-dotenv==1.0.0
# requests==2.31.0
# pydantic==2.5.0

# NEW DEPENDENCE
google-generativeai==0.3.1
google-api-python-client==2.108.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0